+++ doomed_patient/doomed_patient/CellEdge.cpp
@@ -10,7 +10,6 @@ CellEdge::CellEdge(Directions::Direction direction, std::shared_ptr<LevelCell> c
 {
 }
 
-
 CellEdge::~CellEdge()
 {
 }
@@ -25,20 +24,20 @@ void CellEdge::initialisePosition()
 		switch (direction)
 		{
 		case Directions::Direction::NORTH:
-			centreX = cell->getCentreX();
-			centreY = cell->getCentreY() - (cell->getSpriteSizeY() / 2 - sprite->getSpriteSizeY() / 2);
+			centre.x = cell->getCentre().x;
+			centre.y = cell->getCentre().y - (cell->getSpriteSize().y / 2 - sprite->getSpriteSize().y / 2);
 			break;
 		case Directions::Direction::SOUTH:
-			centreX = cell->getCentreX();
-			centreY = cell->getCentreY() + (cell->getSpriteSizeY() / 2 - sprite->getSpriteSizeY() / 2);
+			centre.x = cell->getCentre().x;
+			centre.y = cell->getCentre().y + (cell->getSpriteSize().y / 2 - sprite->getSpriteSize().y / 2);
 			break;
 		case Directions::Direction::EAST:
-			centreX = cell->getCentreX() + (cell->getSpriteSizeX() / 2 - sprite->getSpriteSizeX() / 2);
-			centreY = cell->getCentreY();
+			centre.x = cell->getCentre().x + (cell->getSpriteSize().x / 2 - sprite->getSpriteSize().x / 2);
+			centre.y = cell->getCentre().y;
 			break;
 		case Directions::Direction::WEST:
-			centreX = cell->getCentreX() - (cell->getSpriteSizeX() / 2 - sprite->getSpriteSizeX() / 2);
-			centreY = cell->getCentreY();
+			centre.x = cell->getCentre().x - (cell->getSpriteSize().x / 2 - sprite->getSpriteSize().x / 2);
+			centre.y = cell->getCentre().y;
 			break;
 		}
 	}
@@ -49,6 +48,6 @@ void CellEdge::render(SDL_Renderer* renderer)
 {
 	if (sprite)
 	{
-		sprite->render(renderer, centreX, centreY, sprite->getSpriteSizeX(), sprite->getSpriteSizeY());
+		sprite->render(renderer, centre);
 	}
 }