+++ doomed_patient/doomed_patient/Exit.cpp
@@ -4,12 +4,11 @@
 #include "LevelCell.h"
 
 
-Exit::Exit(PatientGame* game, std::shared_ptr<LevelCell> cell)
+Exit::Exit(PatientGame* game, VectorXY gridCoordinates)
 	: GameObject(game, game->getExitSprite())
 {
 	// Calculate the window position from the grid position
-	centreX = cell->getCentreX();
-	centreY = cell->getCentreY();
+	centre = gridCoordinates.convertGridToWindow();
 }