+++ CHANGELOG.md
@@ -1,20 +1,40 @@
-## 15/03/2016
-* Made multiple doctors and guards spawn in random positions
+##17/03/2016
+* Player now inherits from Character class (Necro)
+* Player starting grid position is now defined by Player (Necro)
+
+##16/03/2016
+* Created VectorXY class (Necro)
+* All classes now use VectorXY for position instead of separate x and y integers (Necro)
+
+##15/03/2016
+* Guards and Doctors now inherit from new Character class (which inherits from GameObject) (Necro)
+* Guards and Doctors now generated using template function (Necro)
+* Made multiple doctors and guards spawn in random positions (Necro)
+
+##11/03/2016
+* Basic player movement now implemented (Necro)
+
+##10/03/2016
+* Player is now created in the bottom left corner (James)
+* A Guard is now created in a random position (James)
+* A Doctor is now created in a random position (James)
 
 ##09/03/2016
-* Exit is now generated on the top or right edge
-* All classes now use smart pointers where appropriate
-* Code in all classes has been tidied up and documented
+* Exit is now generated on the top or right edge (Necro)
+* All classes now use smart pointers where appropriate (Necro)
+* Code in all classes has been tidied up and documented (Necro)
 
 ##07/03/2016
-* Code in Level class has now been tidied up and documented
+* Code in Level class has now been tidied up and documented (Necro)
 
 ##06/03/2016
-* A maze of rooms is now generated using Growing Tree Algorithm
+* A maze of rooms is now generated using Growing Tree Algorithm (Necro)
 
 ##03/03/2016
-* A grid of cells is now generated
+* A grid of cells is now generated (Necro)
 
+##02/03/2016
+* There are now classes for NPCs (James)